+++ Release/ReleaseNotes.md
@@ -1,5 +1,10 @@
 # ReleaseNotes
 
+## [TODO WORK IN PROGRESS] v20.12.28
+- FFT spectrum analyzer functionality. New FV option for FFT vertical scale. T1 and T2 cursors used to select FFT start and end frequency (horizontal scale and offset).
+- Removed bold horizontal line (now signal line is thinner)
+- Some code refactoring
+
 ## v20.12.27
 - Renamed acknowledgement text when saving parameters and screenshot